Problem: Coins are not loaded in same order they are displayed (#798)
Solution: Calculate `quality` field in cryptocurrencies so they can be sorted naturally.
